+++ shared/constants/eventNames.d.ts
@@ -2,6 +2,7 @@ export declare const eventNames: Readonly<{
     APP_INFOS: string;
     WORKING_DIR: string;
     RUN_TEST: string;
+    STOP_COMMAND: string;
     APPROVE_TEST: string;
     INIT_TEST: string;
     RETRIEVE_TEST_RESULT: string;
@@ -42,4 +43,8 @@ export declare const eventNames: Readonly<{
     RETRIEVE_CONFIGURATION_TEMPLATES: string;
     DELETE_CONFIGURATION_TEMPLATE: string;
     RENAME_REFERENCES: string;
+    UPDATE_AVAILABLE: string;
+    UPDATE_DOWNLOADED: string;
+    INSTALL_AND_RESTART: string;
+    DOWNLOAD_UPDATE: string;
 }>;