+++ shared/constants/eventNames.ts
@@ -4,6 +4,7 @@ export const eventNames = Object.freeze({
   RUN_TEST: "runTest",
   APPROVE_TEST: "approveTest",
   INIT_TEST: "initTest",
+  STOP_COMMAND: "stopCommand",
   RETRIEVE_TEST_RESULT: "retrieveTestsResult",
   TEST_LOG: {
     REPLY: "testLog"
@@ -43,5 +44,10 @@ export const eventNames = Object.freeze({
   RETRIEVE_CONFIGURATION_TEMPLATES: "retrieveConfigurationTemplates",
   DELETE_CONFIGURATION_TEMPLATE: "deleteConfigurationTemplate",
 
-  RENAME_REFERENCES: "renameReferences"
+  RENAME_REFERENCES: "renameReferences",
+
+  UPDATE_AVAILABLE: "updateAvailable",
+  UPDATE_DOWNLOADED: "updateDownloaded",
+  INSTALL_AND_RESTART: "installAndRestart",
+  DOWNLOAD_UPDATE: "downloadUpdate"
 });