+++ src/components/app/SaveConfirmationModalComponent.vue
@@ -2,6 +2,7 @@
   <v-dialog
       v-model="dialogDisplayed"
       max-width="400"
+      persistent
     >
     <v-card>
       <v-card-title class="headline">Unsaved changes</v-card-title>