+++ src/components/app/SettingsModalComponent.vue
@@ -6,6 +6,7 @@
       </v-card-title>
       <v-card-text>
         <v-switch :input-value="darkModeEnabled" @change="updateDarkMode($event)" label="Dark mode"/>
+        <v-switch :input-value="autoUpdate" @change="updateAutoUpdate($event)" label="Auto update"/>
       </v-card-text>
       <v-card-actions>
         <v-spacer></v-spacer>
@@ -24,8 +25,14 @@ export default class AboutModalComponent extends Vue {
   @State((state) => state.settingsStore.darkModeEnabled)
   private readonly darkModeEnabled!: boolean;
 
+  @State((state) => state.settingsStore.autoUpdate)
+  private readonly autoUpdate!: boolean;
+
   @Action("settingsStore/updateDarkMode")
   private readonly updateDarkMode!: (newVal: boolean) => Promise<void>;
+
+  @Action("settingsStore/updateAutoUpdate")
+  private readonly updateAutoUpdate!: (newVal: boolean) => Promise<void>;
 
   private dialogDisplayed = true;