+++ src/models/backstopTest.ts
@@ -1,3 +1,4 @@
+import { TestAction } from "./testAction";
 import { Viewport } from "./viewport";
 
 export class BackstopTest {
@@ -22,6 +23,7 @@ export class BackstopTest {
   public misMatchThreshold!: number;
   public requireSameDimensions!: boolean;
   public viewports!: Viewport[];
+  public actions!: TestAction[];
 
   /* for custom fields */
   [key: string]: unknown;