+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## V1.25.0
+### Features
+- Add ability to remove "action" in test view
+- Creating an action has been improved with a dedicated dialog
+- Prevent confirmation modals to be dismissed when clicking outside
+- Add ability to stop test when they are running
+
+### Fixes
+- Fix "help" documentation 
+
+## V1.24.1
+### Fixes
+- Fix path of "action" script
+
+## V1.24.0
+### Features
+- Add "actions", allowing to execute a set of actions such as "focus", "click" etc for your scenario
+- Allowing to enable/disable auto-update 
+- Improve update management by displaying notification
+
+## V1.23.1
+### Fixes
+- fix issue with thread running backstop commands
 ## V1.23.0
 ### Features
 - Add "settings" section and add "dark mode"